requirements-theirs.txt

descartes==1.0.2
Fiona==1.6.0
GDAL==1.11.0
geopandas==0.1.0.dev0
geopy==1.11.0
haversine==0.4.5
imageio==1.5
ipywidgets==5.1.5
iso3166==0.7
matplotlib==1.5.1
Mesa==0.7.5
mplleaflet==0.0.5
nbconvert==4.2.0
networkx==1.11
pprintpp==0.2.3
PySAL==1.11.2
pysd==0.3.2
Rtree==0.8.2
scipy==0.17.1
seaborn==0.7.0
Shapely==1.5.16
six==1.10.0
statsmodels==0.6.1
